getProductPrice and getCartPrice work now

+++ JtProject/src/main/java/com/jtspringproject/JtSpringProject/CartItem.java
@@ -3,10 +3,10 @@ package com.jtspringproject.JtSpringProject;
 public class CartItem {
     private String productName;
     private int quantity;
-    private double totalPrice;
+    private float totalPrice;
     private int productID;
 
-    public CartItem(String productName, int quantity, double totalPrice, int productID) {
+    public CartItem(String productName, int quantity, float totalPrice, int productID) {
         this.productName = productName;
         this.quantity = quantity;
         this.totalPrice = totalPrice;
@@ -19,7 +19,7 @@ public class CartItem {
     public int getQuantity() {
         return quantity;
     }
-    public double getTotalPrice() {
+    public float getTotalPrice() {
         return totalPrice;
     }
     public int getProductID() {return productID;}
